+++ src/Products/PloneMeeting/events.py
@@ -236,18 +236,31 @@ def onMeetingBeforeTransition(meeting, event):
     '''Called before a transition is triggered on a meeting.'''
     # when raising exceptions in a WF script, this needs to be done in the
     # before transition or state is changed nevertheless?
-    if event.new_state.id == 'closed':
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(meeting)
-        if 'return_to_proposing_group' in cfg.getWorkflowAdaptations():
-            # raise a WorkflowException in case there are items still in a
-            # 'returned_to_proposing_group' state
-            returned_to_pg_state_ids = [
-                state for state in cfg.getItemWorkflow(True).states
-                if state.startswith('returned_to_proposing_group')]
+    tool = api.portal.get_tool('portal_plonemeeting')
+    cfg = tool.getMeetingConfig(meeting)
+    wfas = cfg.getWorkflowAdaptations()
+    if event.new_state.id == 'closed' or \
+            (event.new_state.id == 'decisions_published' and
+             'hide_decisions_when_under_writing_check_returned_to_proposing_group' in wfas):
+        # raise a WorkflowException in case there are items returned_to_proposing_group
+        returned_to_pg_state_ids = [
+            state for state in cfg.getItemWorkflow(True).states
+            if state.startswith('returned_to_proposing_group')]
+        if returned_to_pg_state_ids:
             additional_catalog_query = {'review_state': returned_to_pg_state_ids}
-            if meeting.get_items(the_objects=False, additional_catalog_query=additional_catalog_query):
-                msg = _('Can not close a meeting containing items returned to proposing group!')
+            if meeting.get_items(
+                    the_objects=False,
+                    additional_catalog_query=additional_catalog_query):
+                msg = translate(
+                    'Can not set a meeting to ${new_state_title} if it '
+                    'contains items returned to proposing group!',
+                    domain="PloneMeeting",
+                    mapping={
+                        'new_state_title':
+                            translate(cfg.getMeetingWorkflow(True).states[event.new_state.id].title,
+                                      domain="plone",
+                                      context=meeting.REQUEST)},
+                    context=meeting.REQUEST)
                 raise WorkflowException(msg)