+++ CHANGES.rst
@@ -139,6 +139,10 @@ Changelog
   [gbastien]
 - Highlight marginal notes fieldset legend on item view when it contains text.
   [gbastien]
+- Added WFA `hide_decisions_when_under_writing_check_returned_to_proposing_group`
+  that will check that there are no more items `Returned to proposing group` when
+  publishing decisions.
+  [gbastien]
 
 4.2rc34 (2022-09-29)
 --------------------